(fix) do not display token selector when there's only one token available
