+++ app/src/hooks/index.ts
@@ -17,3 +17,4 @@ export { useGraphMarketMakerData } from './useGraphMarketMakerData'
 export { useMarketMakerData } from './useMarketMakerData'
 export { useOutOfBoundsBalance } from './useOutOfBoundsBalance'
 export { usePolling } from './usePolling'
+export { useTokens } from './useTokens'
